Consolidate QKV Parallel Linear slicing and GQA head-replication logic under layers/common
